requirements.txt
transformers>=4.36.0
trl>=0.7.0
peft>=0.7.0
datasets>=2.14.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
torch>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
huggingface_hub>=0.19.0
pandas>=2.0.0
schedule>=1.2.0
